+++ scripts/wait_for_migration_and_extract_keeper_artifacts.sh
@@ -15,10 +15,11 @@ mkdir -p ${CONTRACT_FOLDER}
 
 until [ ${CONTRACTS_READY} -eq 0 ] || [ ${RETRY_COUNT} -eq 120 ]; do
     KEEPER_CONTRACTS_DOCKER_ID=$(docker container ls | grep keeper-contracts | awk '{print $1}')
+    echo "Docker id ${KEEPER_CONTRACTS_DOCKER_ID}"
     docker cp ${KEEPER_CONTRACTS_DOCKER_ID}:/keeper-contracts/artifacts/. ./${CONTRACT_FOLDER}/
     CONTRACT_FILES=(${CONTRACT_FOLDER}/*.json)
     # test for the ready file and a json file
-    if [ -f "${CONTRACT_FILES}" ] && [ -f "${CONTRACT_FOLDER}/dex-ready" ]; then    
+    if [ -f "${CONTRACT_FILES}" ] && [ -f "${CONTRACT_FOLDER}/dex-ready" ]; then
         CONTRACTS_READY=0
         break
     fi
@@ -30,6 +31,20 @@ if [ ${CONTRACTS_READY} -eq 1 ]; then
     echo "Waited for more than ten minutes, but keeper contracts have not been migrated yet. Did you run an Ethereum RPC client and the migration script?"
     exit 1
 fi
+
+if [ -f artifacts/Provenance.development.json ]; then
+    cp $CONTRACT_FOLDER/Provenance.development.json $CONTRACT_FOLDER/Provenance.spree.json
+fi
+
+if [ -f artifacts/DIDRegistry.development.json ]; then
+    cp $CONTRACT_FOLDER/DIDRegistry.development.json $CONTRACT_FOLDER/DIDRegistry.spree.json
+fi
+
+rm $CONTRACT_FOLDER/*.mainnet.json
+rm $CONTRACT_FOLDER/*.pacific.json
+rm $CONTRACT_FOLDER/*.kovan.json
+rm $CONTRACT_FOLDER/*.duero.json
+rm $CONTRACT_FOLDER/*.nile.json
 
 # docker cp ${KEEPER_CONTRACTS_DOCKER_ID}:/keeper-contracts/artifacts/. ./${CONTRACT_FOLDER}/
 echo "copied over the following contracts:"