+++ starfish/d_network/d_network.py
@@ -6,6 +6,9 @@
 """
 
 import logging
+import re
+import time
+import requests
 
 from web3 import (
     HTTPProvider,
@@ -14,7 +17,12 @@ from web3 import (
 from web3.gas_strategies.rpc import rpc_gas_price_strategy
 
 from starfish.contract import ContractManager
-from starfish.exceptions import StarfishInsufficientFunds
+from starfish.exceptions import (
+    StarfishConnectionError,
+    StarfishInsufficientFunds
+)
+from starfish.utils.local_node import get_local_contract_files
+
 
 logger = logging.getLogger(__name__)
 
@@ -42,7 +50,6 @@ CONTRACT_LIST = {
     },
     'DIDRegistry': {
         'name': 'DIDRegistryContract',
-        'abi_filename': 'DIDRegistry.development.json'
     },
     'DirectPurchase': {
         'name': 'DirectPurchaseContract'
@@ -55,7 +62,6 @@ CONTRACT_LIST = {
     },
     'Provenance': {
         'name': 'ProvenanceContract',
-        'abi_filename': 'Provenance.development.json'
     },
 
 }
@@ -68,6 +74,62 @@ class DNetwork():
         self._name = None
         self._contracts = {}
         self._connect(self._url)
+
+    def load_test_node_contracts(self, timeout_seconds=240, sleep_time_seconds=10):
+        """
+
+        This only need to be called on a test network, where the contracts are installed locally on the node.
+
+        Wait for the contracts to be ready and installed
+
+
+        """
+
+        # only do this for the local 'spree' or 'development' node
+
+        if not self._name == 'spree' or self._name == 'development':
+            return True
+
+        # This list order is important, the development contracts have priority over spree
+        test_network_name_list = ['development', 'spree']
+        timeout_time = time.time() + timeout_seconds
+        while timeout_time > time.time():
+            load_items = get_local_contract_files('keeper-contracts', '/keeper-contracts/artifacts')
+
+            # now go through the list and collate the contract artifacts to a dict
+            # we give pereference for .development. contracts
+            contract_items = {}
+
+            for filename, data in load_items.items():
+                if 'name' in data:
+                    match = re.match(r'(\w+)\.(\w+)\.json', filename)
+                    if match:
+                        contract_name = match.group(1)
+                        network_name = match.group(2)
+                        if network_name and network_name in test_network_name_list:
+                            if network_name not in contract_items:
+                                contract_items[network_name] = {}
+                            contract_items[network_name][contract_name] = data
+
+            # now go through the list of contracts supported and load in the artifact data
+            load_count = 0
+            contract_count = 0
+            for contract_name, item in CONTRACT_LIST.items():
+                if item.get('abi_filename', True):
+                    contract_count += 1
+                    for network_name in test_network_name_list:
+                        if network_name in contract_items and contract_name in contract_items[network_name]:
+                            data = contract_items[network_name][contract_name]
+                            self._contract_manager.set_contract_data('spree', contract_name, data)
+                            logger.debug(f'imported contract {contract_name}.{network_name}')
+                            load_count += 1
+                            break
+            if load_count == contract_count:
+                return True
+            # take some sleep to wait for the contracts to be built
+            logger.debug(f'only loaded {load_count} out of {contract_count} contracts wating for local node to startup..')
+            time.sleep(sleep_time_seconds)
+        return False
 
     def get_contract(self, name):
         if name not in CONTRACT_LIST:
@@ -224,7 +286,11 @@ class DNetwork():
         self._url = url
         self._web3 = Web3(HTTPProvider(url))
         if self._web3:
-            self._name = DNetwork.find_network_name_from_id(int(self._web3.net.version))
+            try:
+                self._name = DNetwork.find_network_name_from_id(int(self._web3.net.version))
+            except requests.exceptions.ConnectionError as e:
+                raise StarfishConnectionError(e)
+
             logger.info(f'connected to the {self._name} network')
             self._web3.eth.setGasPriceStrategy(rpc_gas_price_strategy)
             self._contract_manager = ContractManager(self._web3, DEFAULT_PACKAGE_NAME)