+++ starfish/exceptions.py
@@ -5,6 +5,10 @@
 """
 
 
+class StarfishConnectionError(Exception):
+    """ Raised when starfish cannot connect to the network """
+
+
 class StarfishPurchaseError(Exception):
     """ Raised when a purchase events have failed to complete """